feat: change default hijack to only "development"
BREAKING CHANGE: The default for importing the hijack file has changed from `!== 'production'` to `=== 'development'`. This way, in addition to not hijacking "production" it doesn't hijack for "test" either.

If you prefer to keep the original behavior, then you can make this change after upgrading:

```diff
- import 'stop-runaway-react-effects/hijack'
+ import {hijackEffects} from 'stop-runaway-react-effects'

+ if (process.env.NODE_ENV !== 'production') hijackEffects()
```

But it is recommended to just stick with the default and not hijack for tests.